recontent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg[binary]==3.2.1
pg8000==1.31.2
cloud-sql-python-connector==1.9.1
python-dotenv==1.0.1
stripe==9.11.0
google-cloud-storage==2.18.2
google-cloud-pubsub==2.22.0
google-cloud-secret-manager==2.20.2
google-cloud-logging==3.11.2
google-cloud-aiplatform==1.65.0
Pillow==10.4.0
tenacity==8.5.0
